refactor: Split tests into integration and unit tests
