+++ tests/integration/config.py
@@ -1,0 +1,5 @@
+from pathlib import Path
+
+EXPECT_DATA_PATH = Path(__file__).parent / "expectdata"
+EXPECT_CONFIG_PATH = EXPECT_DATA_PATH / "config"
+EXPECT_STATEMENTS_PATH = EXPECT_DATA_PATH / "statements"
