+++ tests/integration/expectdata/config/default-cat-annual-stockrow.json
