+++ tests/integration/expectdata/config/no-adjust-cat-annual-capiq.json
