+++ tests/integration/expectdata/config/no-adjust-cat-annual-stockrow.json
