+++ tests/integration/expectdata/statements/fcst_capiq_cat_annual.py
