+++ tests/integration/expectdata/statements/fcst_capiq_cat_quarterly.py
