+++ tests/integration/expectdata/statements/fcst_stockrow_cat_bs_diff_100000_annual.py
