+++ tests/integration/expectdata/statements/fcst_stockrow_cat_no_balance_annual.py
