+++ tests/integration/expectdata/statements/fcst_stockrow_cat_plug_make_forecast_annual.py
