+++ tests/integration/expectdata/statements/fcst_stockrow_cat_quarterly.py
