+++ tests/integration/expectdata/statements/fcst_stockrow_mar_annual.py
