+++ tests/integration/expectdata/statements/fcst_stockrow_mar_quarterly.py
