+++ tests/integration/expectdata/statements/load_capiq_cat_annual.py
