+++ tests/integration/expectdata/statements/load_stockrow_cat_quarterly.py
