+++ tests/integration/expectdata/statements/load_stockrow_mar_annual.py
