+++ tests/integration/expectdata/statements/load_stockrow_mar_quarterly.py
