+++ tests/integration/test_config.py
@@ -3,7 +3,8 @@ from pathlib import Path
 
 from finstmt import BalanceSheets, FinancialStatements, IncomeStatements
 from finstmt.config_manage.statements import StatementsConfigManager
-from tests.conftest import DEVELOPMENT_MODE, EXPECT_CONFIG_PATH
+from tests.conftest import DEVELOPMENT_MODE
+from tests.integration.config import EXPECT_CONFIG_PATH
 
 
 def check_config(config: StatementsConfigManager, json_path: Path):
@@ -62,17 +63,3 @@ def test_config_load_cat_annual_capiq_no_adjust(
     )
     json_path = EXPECT_CONFIG_PATH / "no-adjust-cat-annual-capiq.json"
     check_config(stmts.config, json_path)
-
-
-def test_config_attribute_access(
-    annual_capiq_income_stmt: IncomeStatements, annual_capiq_bs_stmt: BalanceSheets
-):
-    dates = annual_capiq_income_stmt.dates
-    stmts = FinancialStatements(annual_capiq_income_stmt, annual_capiq_bs_stmt[dates])
-    # Check reading a config value by attribute access
-    assert stmts.config.cash.display_name == "Cash and Cash Equivalents"
-
-    # Check updating a config value by attribute access
-    stmts.config.cash.display_name = "Cash"
-    assert stmts.config.cash.display_name == "Cash"
-    assert "cash" in dir(stmts.config)
