--- tests/test_placeholder.py
@@ -1,3 +1,0 @@
-# TODO: Add real tests
-def test_placeholder():
-    assert True
